feat: support multi version nav in mobile mode
