+++ packages/core/src/theme-default/components/NavScreen/index.module.scss
@@ -44,3 +44,45 @@
 .socialAndAppearance {
   padding: 8px 0 7px;
 }
+
+.navScreenMenuGroup {
+  height: 39px;
+  overflow: hidden;
+  transition: border-color 0.5s;
+}
+.navScreenMenuGroup.open {
+  padding-bottom: 10px;
+  height: auto;
+}
+.button {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 8px 0 7px 0;
+  width: 100%;
+  line-height: 24px;
+  font-size: 14px;
+  font-weight: 500;
+  color: var(--rp-c-text-1);
+  transition: color 0.25s;
+}
+
+.buttonSpan {
+  display: flex;
+  align-items: center;
+  height: 24px;
+  padding-right: 4px;
+}
+.navScreenMenuGroup .items {
+  visibility: hidden;
+}
+
+.navScreenMenuGroup.open .items {
+  visibility: visible;
+}
+.down {
+  transition: all 0.5s;
+}
+.open.down {
+  transform: rotate(180deg);
+}
